+++ home/yzhao/code-retreat-20240126/bard/main_bard.go
@@ -1,4 +1,4 @@
-package main
+package gencode
 
 import (
     "fmt"
