+++ home/yzhao/code-retreat-20240126/gpt4/main_test.go
@@ -1,4 +1,4 @@
-package main
+package gpt4
 
 import (
     "reflect"
